separate the arena modifier from spot driver
