+++ setup.py
@@ -50,6 +50,7 @@ setup(
             "set_initial_pose = " + package_name + ".set_initial_pose:main",
             "gpp_stacker = " + package_name + ".gpp_stacker:main",
             "retract_manipulator = " + package_name + ".retract_manipulator:main",
+            "arena_modifier = " + package_name + ".arena_modifier:main",
         ],
     },
 )
